+++ lib/folio/engine.rb
@@ -42,6 +42,10 @@ module Folio
     config.folio_console_clonable_enabled = true
     config.folio_console_audited_revisions_limit = 50
     config.folio_console_preview_url_for_procs = nil
+    config.folio_console_files_thumbnail_groups_proc = -> (groups:, site:) { groups }
+    config.folio_console_files_thumbnail_groups_main_ratio_proc = lambda do |ratio:, **|
+      Folio::Console::Files::ThumbnailGroups::MainRatioGroups.default_ratio(ratio:)
+    end
 
     config.folio_rewriter_lambda_for_has_console_url = nil